# Include ugu for all course offerings cache