--- a/dashboard/test/models/course_offering_test.rb
+++ b/dashboard/test/models/course_offering_test.rb
@@ -605,11 +605,11 @@ class CourseOfferingTest < ActiveSupport::TestCase
   test 'query count for assignable_course_offerings' do
     Unit.stubs(:should_cache?).returns true
 
-    assert_cached_queries(16) do
+    assert_cached_queries(0) do
       CourseOffering.assignable_course_offerings_info(@teacher)
     end
 
-    assert_cached_queries(18) do
+    assert_cached_queries(0) do
       CourseOffering.assignable_course_offerings_info(@facilitator)
     end