Since the standalone unit migration, users have experienced extended loading times on the My Courses table:
<img width="1039" height="652" alt="image" src="https://github.com/user-attachments/assets/91309771-9f8c-4378-9b3b-8cf6ac935ece" />

While updating tests to remove references to standalone units, we discovered that the query count for `assignable_course_offerings_info` increased substantially (#66714 and #65961). This PR now includes unit group units when it loads the course offering cache. 

## Links
- Jira: [TEACH-2067](https://codedotorg.atlassian.net/browse/TEACH-2067)

## Testing story
Loaded the cache change to an adhoc to compare timings of a few key calls:
`quick_assign_course_offerings`: load the course offerings when creating a section
`valid_course_offerings`: load the course in the owned section table

|                                 | Production | Adhoc - Before | Adhoc - After |
|---------------------------------|------------|----------------|---------------|
| `quick_assign_course_offerings` | 1.83s      | 4.2s           | ~1.5s         |
| `valid_course_offerings`        | 2.11s      | 12s            | ~1.5s         |


## Follow-up work
Once this change makes it into production, I will see if the timing improves at all. If not, then we will need to open another ticket to re-investigate.